Fallback fix in event of 0,0 gamma/boost
Bordercities will now grab the default values for the Gamma and Boost
per session and use them as a fallback in the event that the previously
stored 'oldGamma' and 'oldBoost' are saved as 0,0.  I am going to be
working on this particular implementation to iron out any redundancy
later, however, it performs as intended for the moment.