Repurposed new 'Overdrive' mode as highlight
Loving the 'Sobelcities:Overdrive.'  Renamed it to '++' for simplicity
and am highlighting it due to its ability to achieve the
exaggerated-edges effect as strongly as it does AND as cleanly as it
does.  Zoomed out edges have always been an eye-sore, and today, much
less so.  Work on distance edge quality doesn't end here - there will be
further tweaks which I hope will render this current 'best' preset as
looking terrible (in comparison).  I'm not going to be satisfied until
the edge detection looks as it ought to at max-distance zoom.